Add preloadedImageComposable
