+++ src/components/feature-board-file-element/composables/preloadedImage.composable.ts
@@ -1,0 +1,17 @@
+import { onMounted, ref } from "vue";
+
+export const usePreloadedImage = (imageUrl: string) => {
+	const isImageLoading = ref(true);
+
+	onMounted(() => {
+		const img = new Image();
+		img.src = imageUrl;
+		img.onload = () => {
+			isImageLoading.value = false;
+		};
+	});
+
+	return {
+		isImageLoading,
+	};
+};
